student-analyzer/requirements.txt
flask
pandas
openpyxl
google-generativeai